chore(tools): relocate minimax inference profile
