+++ tools/minimax_music3_worker.py
@@ -18,7 +18,7 @@ PROJECT_ROOT = Path(__file__).resolve().parents[1]
 if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
 
-from minimax_profile import (  # noqa: E402
+from tools.minimax_profile import (  # noqa: E402
     MiniMaxInferenceProfile,
     MiniMaxProfileError,
     minimax_profile_from_mapping,
