+++ tools/verify_b02_scope.py
@@ -44,6 +44,7 @@ B02_CONTRACT_PATHS = frozenset(
         "music_duration.py",
         "music_renderer.py",
         "music_reply.py",
+        "minimax_profile.py",
         "patch_feapp.py",
         "pyproject.toml",
         "reply_delivery.py",
@@ -58,6 +59,7 @@ B02_CONTRACT_PATHS = frozenset(
         "requirements-ci.txt",
         "requirements-dev.txt",
         "tools/minimax_music3_worker.py",
+        "tools/minimax_profile.py",
         "tools/Install-ThirdParty.ps1",
         "tools/verify_b02_scope.py",
         "tts/delivery.py",
@@ -125,6 +127,7 @@ ALLOWED_MUTATIONS = frozenset(
         "music_duration.py",
         "music_renderer.py",
         "music_reply.py",
+        "minimax_profile.py",
         "patch_feapp.py",
         "pyproject.toml",
         "reply_delivery.py",
@@ -147,6 +150,7 @@ ALLOWED_MUTATIONS = frozenset(
         "tests/packaging/test_B10A_runtime.py",
         "runtime/packaging/b10a/manager.py",
         "tools/minimax_music3_worker.py",
+        "tools/minimax_profile.py",
         "tools/Install-ThirdParty.ps1",
         "tools/verify_project_status.py",
         "tts/delivery.py",
